rag_lab/server/requirements-dev.txt
# ================================================================
# RAG Lab Server 개발 의존성
# ================================================================

# 기본 의존성 포함
-r requirements.txt

# ================================================================
# 코드 품질 도구
# ================================================================

# 린터 및 포매터
ruff==0.5.1                    # 빠른 Python 린터
black==24.4.2                  # 코드 포매터
isort==5.13.2                  # import 정렬

# 타입 체킹
mypy==1.10.1                   # 정적 타입 검사기
types-requests==2.32.0.20240622

# ================================================================
# 테스트 도구
# ================================================================

# 테스트 프레임워크
pytest==8.2.2                 # 테스트 프레임워크
pytest-asyncio==0.23.7        # 비동기 테스트 지원
pytest-cov==5.0.0             # 커버리지 측정
pytest-mock==3.14.0           # 목 객체 지원

# API 테스트
httpx==0.27.0                  # 비동기 HTTP 클라이언트
pytest-httpx==0.30.0          # httpx용 pytest 플러그인

# ================================================================
# 문서화 도구
# ================================================================

# 문서 생성
mkdocs==1.6.0                 # 문서 사이트 생성
mkdocs-material==9.5.27       # Material 테마
mkdocs-swagger-ui-tag==0.6.10 # API 문서 통합

# Jupyter 지원
jupyter==1.1.1                # Jupyter Notebook
ipykernel==6.29.5             # IPython 커널

# ================================================================
# 개발 편의 도구
# ================================================================

# 가상환경 관리
pip-tools==7.4.1              # requirements 관리

# 디버깅
ipdb==0.13.13                 # 향상된 디버거
rich==14.0.0                  # 색상 출력

# 성능 프로파일링
line-profiler==4.1.3          # 라인별 성능 측정
memory-profiler==0.61.0       # 메모리 사용량 측정

# ================================================================
# 보안 도구
# ================================================================

# 보안 취약점 검사
safety==3.2.4                 # 취약점 데이터베이스 검사
bandit==1.7.9                 # 보안 이슈 검사

# ================================================================
# CI/CD 도구
# ================================================================

# 사전 커밋 훅
pre-commit==3.7.1             # Git 훅 관리

# 환경 검증
python-dotenv==1.0.1          # 환경변수 관리
pydantic==2.11.7               # 데이터 검증

# ================================================================
# 개발 서버 도구
# ================================================================

# 핫 리로드
watchdog==4.0.1               # 파일 변경 감지

# API 개발
fastapi-cli==0.0.4            # FastAPI CLI 도구